tests/wpt/web-platform-tests/css/selectors/focus-in-focusin-event-001.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Selectors Level 4: :focus, :focus-visible and :focus-within in focusin event</title>
<link rel="help" href="https://drafts.csswg.org/selectors-4/#focus-pseudo">
<link rel="help" href="https://drafts.csswg.org/selectors-4/#focus-visible-pseudo">
<link rel="help" href="https://drafts.csswg.org/selectors-4/#focus-within-pseudo">
<link rel="help" href="https://w3c.github.io/uievents/#event-type-focusin">
<meta name='author' title='Manuel Rego Casasnovas' href='mailto:rego@igalia.com'>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<input type="text">
<script>
var input = document.querySelector('input');
input.addEventListener('focusin', function(e) {
  var focusPseudo = document.querySelector(':focus');
  var focusVisiblePseudo = document.querySelector(':focus-visible');
  var focusWithinPseudo = document.querySelector(':focus-within');
  test(() => {
    assert_equals(e.target, focusPseudo, "':focus' matches event.target");
  }, "Checks that ':focus' pseudo-class matches inside 'focusin' event handler");
  test(() => {
    assert_equals(e.target, focusVisiblePseudo, "':focus-visible' matches event.target");
  }, "Checks that ':focus-visible' pseudo-class matches inside 'focusin' event handler");
  test(() => {
    assert_equals(document.documentElement, focusWithinPseudo, "':focus-within' matches document.documentElement");
  }, "Checks that ':focus-within' pseudo-class matches inside 'focusin' event handler");
}, false);
input.focus();
</script>
